remove redundant codes
