+++ notebooks/run_ufconf_denoise.py
@@ -351,7 +351,6 @@ def main(args):
         - device (str): The device to run the model on (e.g., 'cuda:0').
         - chunk_size (str): Chunk size for processing.
         - model (str): Name of the ufconf model to use.
-        - data_path (str): Path to the dataset containing PDB and MSA data.
         - data_idx (str): Index to specify the random seed for MSA.
         - from_pdb (bool): Flag to indicate whether to run inference with on-the-fly generated MSA.
         - bf16 (bool): Flag to indicate the use of bfloat16 precision.
@@ -393,7 +392,7 @@ def main(args):
             for attempt in range(max_retries):
                 try:
                     # Attempt to load the features
-                    feat_raw, all_chain_labels = utils.load_features_from_pdb(args, config, Job, dir_feat_name, cif=True)
+                    feat_raw, all_chain_labels = utils.load_features_from_pdb(args, Job, dir_feat_name, cif=True)
                     break  # If successful, exit the loop
                 except Exception as e:  # Catching a general exception
                     if attempt < max_retries - 1:
@@ -408,7 +407,7 @@ def main(args):
             for attempt in range(max_retries):
                 try:
                     # Attempt to load the features
-                    feat_raw, all_chain_labels = utils.load_features_from_pdb(args, config, Job, dir_feat_name, cif=False)
+                    feat_raw, all_chain_labels = utils.load_features_from_pdb(args, Job, dir_feat_name, cif=False)
                     break  # If successful, exit the loop
                 except Exception as e:  # Catching a general exception
                     if attempt < max_retries - 1:
@@ -420,8 +419,7 @@ def main(args):
                         print("Maximum retries reached. Aborting.")
                         raise
         else:
-            # use existing PDB and MSA datasets to get all the features and labels for query Protein ID
-            feat_raw, all_chain_labels = utils.load_features_from_lmdb(args, config, Job, job_name)
+            print(f"Please provide valid flag: 'from_pdb' or 'from_cif'.")
         # print("feat raw keys", feat_raw.keys())
         # print("feat raw msa", feat_raw["msa"].shape)
         # print("feat raw 10", feat_raw["msa"][:10])
@@ -548,10 +546,6 @@ The keyword arguments are illustrated in ufconf/README.md.
     parser.add_argument(
         "--model", type=str, default="ufconf_af2_v3_c",
         help="the name of the model. (default: `ufconf_af2_v3_c`)."
-    )
-    parser.add_argument(
-        "--data_path", type=str, default="/mnt/vepfs/fs_projects/unifold/data_0916/traineval/",
-        help="the path for the downloaded PDB and MSA dataset"
     )
     parser.add_argument(
         "--data_idx", type=str, default="0",
