+++ ufconf/__init__.py
@@ -1,3 +1,3 @@
 """ Diffold: Uni-Fold x Diffusion."""
 
-from . import task, model, loss, dataset
+from . import model, loss
