+++ unifold/config.py
@@ -2,6 +2,7 @@ import ml_collections as mlc
 import copy
 from typing import Any
 
+
 N_RES = "number of residues"
 N_MSA = "number of MSA sequences"
 N_EXTRA_MSA = "number of extra MSA sequences"
@@ -110,14 +111,12 @@ def base_config():
                     "random_delete_msa": {
                         "max_msa_entry": 1 << 25,  # := 33554432
                     },
-                    "use_msa": True,
                     "v2_feature": False,
                     "gumbel_sample": False,
                     "max_extra_msa": 1024,
                     "msa_cluster_features": True,
                     "reduce_msa_clusters_by_max_templates": True,
                     "resample_msa_in_recycling": True,
-                    "train_max_date": "2022-04-30",
                     "template_features": [
                         "template_all_atom_positions",
                         "template_sum_probs",
@@ -198,8 +197,8 @@ def base_config():
                     "max_msa_clusters": 128,
                     "max_templates": 4,
                     "num_ensembles": 1,
-                    "crop": True,
-                    "crop_size": 384,
+                    "crop": False,
+                    "crop_size": None,
                     "spatial_crop_prob": 0.5,
                     "ca_ca_threshold": 10.0,
                     "supervised": True,
@@ -364,13 +363,11 @@ def base_config():
                         "num_bins": 50,
                         "d_in": d_single,
                         "d_hid": 128,
-                        "enabled": True,
                     },
                     "distogram": {
                         "d_pair": d_pair,
                         "num_bins": aux_distogram_bins,
                         "disable_enhance_head": False,
-                        "enabled": True,
                     },
                     "pae": {
                         "d_pair": d_pair,
@@ -383,7 +380,6 @@ def base_config():
                         "d_msa": d_msa,
                         "d_out": 23,
                         "disable_enhance_head": False,
-                        "enabled": True,
                     },
                     "experimentally_resolved": {
                         "d_single": d_single,
@@ -555,28 +551,12 @@ def model_config(name, train=False):
         c.globals.alphafold_original_mode = True
     elif name == "model_2_v2":
         c = model_2_v2(c)
-    elif name == "model_1_v2_ft":
-        c = model_2_v2(c)
-        recursive_set(c, "max_extra_msa", 5120)
-        recursive_set(c, "max_msa_clusters", 512)
-        c.data.train.crop_size = 384
-        c.loss.violation.weight = 0.02
     elif name == "model_2_v2_ft":
         c = model_2_v2(c)
         recursive_set(c, "max_extra_msa", 1024)
         recursive_set(c, "max_msa_clusters", 512)
         c.data.train.crop_size = 384
         c.loss.violation.weight = 0.02
-    elif name == "model_3_v2_ft" or name == "model_4_v2_ft":
-        c = model_2_v2(c)
-        recursive_set(c, "max_extra_msa", 5120)
-        recursive_set(c, "max_msa_clusters", 512)
-        c.data.train.crop_size = 384
-        c.loss.violation.weight = 0.02
-        c.model.template.enabled = False
-        c.model.template.embed_angles = False
-        recursive_set(c, "use_templates", False)
-        recursive_set(c, "use_template_torsion_angles", False)
     elif name == "model_3_af2" or name == "model_4_af2":
         recursive_set(c, "max_extra_msa", 5120)
         recursive_set(c, "max_msa_clusters", 512)
@@ -586,16 +566,6 @@ def model_config(name, train=False):
         c.model.heads.experimentally_resolved.enabled = True
         c.loss.experimentally_resolved.weight = 0.01
         c.globals.alphafold_original_mode = True
-        c.model.template.enabled = False
-        c.model.template.embed_angles = False
-        recursive_set(c, "use_templates", False)
-        recursive_set(c, "use_template_torsion_angles", False)
-    elif name == "model_5_v2_ft":
-        c = model_2_v2(c)
-        recursive_set(c, "max_extra_msa", 1024)
-        recursive_set(c, "max_msa_clusters", 512)
-        c.data.train.crop_size = 384
-        c.loss.violation.weight = 0.02
         c.model.template.enabled = False
         c.model.template.embed_angles = False
         recursive_set(c, "use_templates", False)
